requirements.txt
pywebview
cefpython3; sys_platform == "win32"
